Number of establishments by establishment size.
This slide showcases an interactive dashboard for businesses in Saudi Arabia in 2023, developed using Power BI. Statistics include the number of businesses by size (micro, small, medium, and large), as well as their geographical distribution. It also illustrates the tasks performed using Power Query and the implementation of dashboard elements, such as cards and interactive indicators, to enhance data analysis and decision-making.